Refactor helpers
