+++ app/controllers/web/application_controller.rb
@@ -3,7 +3,6 @@
 class Web::ApplicationController < ApplicationController
   protect_from_forgery
 
-  include AuthHelper
   include Flash
   include Pundit::Authorization
