+++ app/helpers/application_helper.rb
@@ -1,7 +1,4 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  include ViewHelper
-  include UsersHelper
-  include CategoriesHelper
 end
