+++ app/views/layouts/shared/_header.html.slim
@@ -13,7 +13,7 @@ header.px-3.d-flex *{ class: %w[flex-wrap align-items-center \
           .dropdown-menu
             - all_categories.each do |category|
               = link_to category.title, category_path(category), class: 'dropdown-item'
-      - if user_signed_in?
+      - if signed_in?
         - if current_user.admin?
           == render partial: 'layouts/shared/menu_admin'
         - else
